+++ MystatDesktopWpf/UserControls/Menus/Debug.xaml
