+++ MystatDesktopWpf/UserControls/Menus/Debug.xaml.cs
