+++ MystatDesktopWpf/UserControls/Menus/MainMenu.xaml.cs
@@ -14,8 +14,8 @@ using System.Windows.Navigation;
 using System.Windows.Shapes;
 using MaterialDesignThemes.Wpf;
 using MaterialDesignThemes.Wpf.Transitions;
-using MystatDesktopWpf.Domain;
 using MystatDesktopWpf.Services;
+using MystatDesktopWpf.ViewModels;
 
 namespace MystatDesktopWpf.UserControls
 {
