+++ MystatDesktopWpf/UserControls/Menus/Placeholder.xaml
