+++ MystatDesktopWpf/UserControls/Menus/Placeholder.xaml.cs
