+++ MystatDesktopWpf/UserControls/Menus/Schedule.xaml.cs
