+++ MystatDesktopWpf/UserControls/Menus/Settings.xaml
@@ -1,0 +1,23 @@
+﻿<UserControl x:Class="MystatDesktopWpf.UserControls.Settings"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:MystatDesktopWpf"
+             xmlns:domain="clr-namespace:MystatDesktopWpf.Domain"
+             xmlns:controls="clr-namespace:MystatDesktopWpf.UserControls"
+             xmlns:materialDesign="http://materialdesigninxaml.net/winfx/xaml/themes"
+             xmlns:vm="clr-namespace:MystatDesktopWpf.ViewModels"
+             xmlns:converters="clr-namespace:MystatDesktopWpf.Converters"
+             mc:Ignorable="d" 
+             d:DesignHeight="450" d:DesignWidth="800"
+             Background="{DynamicResource MaterialDesignPaper}">
+
+    <ScrollViewer VerticalScrollBarVisibility="Auto">
+        <StackPanel Margin="16">
+            <controls:NotificationSettings Margin="8"/>
+            <controls:ThemeSettings Margin="8"/>
+        </StackPanel>
+    </ScrollViewer>
+    
+</UserControl>
