+++ MystatDesktopWpf/UserControls/Menus/Settings.xaml.cs
@@ -1,0 +1,33 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Text.RegularExpressions;
+using System.Threading.Tasks;
+using System.Windows;
+using System.Windows.Controls;
+using System.Windows.Data;
+using System.Windows.Documents;
+using System.Windows.Input;
+using System.Windows.Media;
+using System.Windows.Media.Imaging;
+using System.Windows.Navigation;
+using System.Windows.Shapes;
+using MaterialDesignThemes.Wpf;
+using MystatDesktopWpf.Services;
+using System.ComponentModel;
+using MystatDesktopWpf.ViewModels;
+
+namespace MystatDesktopWpf.UserControls
+{
+    /// <summary>
+    /// Interaction logic for Settings.xaml
+    /// </summary>
+    public partial class Settings : UserControl
+    {
+        public Settings()
+        {
+            InitializeComponent();
+        }
+    }
+}
