+++ MystatDesktopWpf/UserControls/NotificationSettings.xaml
@@ -1,0 +1,87 @@
+﻿<UserControl x:Class="MystatDesktopWpf.UserControls.NotificationSettings"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:materialDesign="http://materialdesigninxaml.net/winfx/xaml/themes"
+             xmlns:vm="clr-namespace:MystatDesktopWpf.ViewModels"
+             xmlns:domain="clr-namespace:MystatDesktopWpf.Domain"
+             xmlns:converters="clr-namespace:MystatDesktopWpf.Converters"
+             xmlns:local="clr-namespace:MystatDesktopWpf.UserControls" 
+             d:DataContext="{d:DesignInstance Type=vm:NotificationSettingsViewModel}"
+             mc:Ignorable="d" 
+             d:DesignHeight="450" d:DesignWidth="800">
+    
+    <UserControl.Resources>
+        <converters:ComparisonConverter x:Key="ComparisonConverter"/>
+        <vm:NotificationSettingsViewModel x:Key="NotificationSettingsViewModel"/>
+    </UserControl.Resources>
+    
+    <materialDesign:Card>
+        <Grid>
+            <Grid.ColumnDefinitions>
+                <ColumnDefinition/>
+            </Grid.ColumnDefinitions>
+            <Grid.RowDefinitions>
+                <RowDefinition Height="Auto"/>
+                <RowDefinition Height="Auto"/>
+                <RowDefinition Height="Auto"/>
+            </Grid.RowDefinitions>
+
+            <TextBlock Text="Напоминания о начале пары" HorizontalAlignment="Left"
+                           Style="{DynamicResource MaterialDesignBody1TextBlock}" Margin="8"/>
+            <ToggleButton HorizontalAlignment="Right" Grid.Column="1" Style="{StaticResource MaterialDesignSwitchToggleButton}"
+                              IsChecked="{Binding NotificationEnabled}" x:Name="notifierCheckBox" Margin="8"/>
+
+            <WrapPanel Grid.Row="1">
+                <StackPanel IsEnabled="{Binding ElementName=notifierCheckBox, Path=IsChecked}" Margin="8"
+                            Grid.Row="1">
+                    <Label Content="Какие пары напоминать:" FontSize="14" Margin="0,8,0,0"/>
+                    <RadioButton GroupName="scheduleType" Content="Только первую пару" Margin="0,0,0,8" IsChecked="{Binding OnlyFirstSchedule}"/>
+                    <RadioButton GroupName="scheduleType" Content="Каждую пару" Margin="0,0,0,8" IsChecked="{Binding IsNotOnlyFirstSchedule}"/>
+                </StackPanel>
+                <Rectangle VerticalAlignment="Stretch" Fill="{DynamicResource MaterialDesignBody}" Width="1" Margin="8"/>
+
+                <StackPanel IsEnabled="{Binding ElementName=notifierCheckBox, Path=IsChecked}" Margin="8"
+                            Grid.Row="1" Grid.Column="1">
+                    <Label Content="Когда напоминать:" FontSize="14" Margin="0,8,0,0"/>
+
+                    <RadioButton IsChecked="{Binding Path=DelayMode, Converter={StaticResource ComparisonConverter}, ConverterParameter={x:Static domain:NotificationDelayMode.WithoutDelay}}"
+                                     Content="Напоминать в момент начала" Margin="0,0,0,8" GroupName="DelayMode"/>
+
+                    <StackPanel Orientation="Horizontal" Grid.Row="1" Grid.RowSpan="2" VerticalAlignment="Center">
+                        <RadioButton IsChecked="{Binding Path=DelayMode, Converter={StaticResource ComparisonConverter}, ConverterParameter={x:Static domain:NotificationDelayMode.Delayed}}"
+                                         Content="Напоминать заранее за" Margin="0, 0, 0, 8" Grid.Row="1" GroupName="DelayMode"/>
+                        <TextBox MaxLength="2" Margin="8, 0, 8, 0" Width="20" HorizontalContentAlignment="Center"
+                                     materialDesign:TextFieldAssist.CharacterCounterVisibility="Collapsed" x:Name="minutesTextBox"
+                                     Text="{Binding NotificationDelay, Mode=OneTime}" PreviewTextInput="TextBox_PreviewTextInput" TextChanged="minutesTextBox_TextChanged" />
+                        <Label x:Name="minutesLabel" Content="минут"/>
+                    </StackPanel>
+                    <RadioButton IsChecked="{Binding Path=DelayMode, Converter={StaticResource ComparisonConverter}, ConverterParameter={x:Static domain:NotificationDelayMode.Both}}"
+                             Content="Оба варианта" Margin="0,0,0,8" Grid.Row="2" GroupName="DelayMode"/>
+
+
+                </StackPanel>
+                <Rectangle VerticalAlignment="Stretch" Fill="{DynamicResource MaterialDesignBody}" Width="1" Margin="8"/>
+
+                <Grid Grid.Row="2" Grid.ColumnSpan="2" Margin="8"
+                            IsEnabled="{Binding ElementName=notifierCheckBox, Path=IsChecked}">
+                    <Grid.RowDefinitions>
+                        <RowDefinition Height="Auto"/>
+                        <RowDefinition/>
+                    </Grid.RowDefinitions>
+                    <Grid.ColumnDefinitions>
+                        <ColumnDefinition Width="Auto"/>
+                        <ColumnDefinition/>
+                    </Grid.ColumnDefinitions>
+                    <Label Content="Звук уведомления:" FontSize="14" Margin="0,8,0,0" HorizontalAlignment="Center" Grid.ColumnSpan="2"/>
+                    <Slider Margin="8,8,8,8" VerticalAlignment="Center" Grid.Row="1" x:Name="notificationVolumeSlider"
+                                Grid.Column="1" Value="{Binding NotificationVolume}" Maximum="1"/>
+                    <Button x:Name="muteButton" Grid.Row="1" Grid.Column="0" Style="{DynamicResource MaterialDesignIconButton}" Width="24" Height="24" Click="Button_Mute_Click">
+                        <materialDesign:PackIcon x:Name="muteIcon" Kind="Mute" Foreground="{DynamicResource MaterialDesignCheckBoxOff}"/>
+                    </Button>
+                </Grid>
+            </WrapPanel>
+        </Grid>
+    </materialDesign:Card>
+</UserControl>
