+++ MystatDesktopWpf/UserControls/NotificationSettings.xaml.cs
@@ -1,5 +1,8 @@
-﻿using System;
+﻿using MaterialDesignThemes.Wpf;
+using MystatDesktopWpf.ViewModels;
+using System;
 using System.Collections.Generic;
+using System.ComponentModel;
 using System.Linq;
 using System.Text;
 using System.Text.RegularExpressions;
@@ -13,29 +16,27 @@ using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Navigation;
 using System.Windows.Shapes;
-using MaterialDesignThemes.Wpf;
-using MystatDesktopWpf.Domain;
-using MystatDesktopWpf.Services;
-using System.ComponentModel;
 
 namespace MystatDesktopWpf.UserControls
 {
     /// <summary>
-    /// Interaction logic for Settings.xaml
+    /// Interaction logic for NotificationSettings.xaml
     /// </summary>
-    public partial class Settings : UserControl
+    public partial class NotificationSettings : UserControl
     {
-        SettingsViewModel viewModel;
+        NotificationSettingsViewModel viewModel;
         double savedVolume = 1;
-        public Settings()
+
+        public NotificationSettings()
         {
+            viewModel = new NotificationSettingsViewModel();
+            DataContext = viewModel;
+
             InitializeComponent();
-            viewModel = (SettingsViewModel)FindResource("SettingsViewModel");
-            
+
             viewModel.PropertyChanged += ViewModel_PropertyChanged;
             ViewModel_PropertyChanged(null, new PropertyChangedEventArgs("NotificationVolume"));
         }
-
         private void ViewModel_PropertyChanged(object? sender, PropertyChangedEventArgs e)
         {
             if (e.PropertyName == "NotificationVolume")
@@ -56,10 +57,9 @@ namespace MystatDesktopWpf.UserControls
 
         private void minutesTextBox_TextChanged(object sender, TextChangedEventArgs e)
         {
-            SettingsViewModel settings = (SettingsViewModel)this.FindResource("SettingsViewModel");
             int result;
             if (int.TryParse(minutesTextBox.Text, out result))
-                settings.NotificationDelay = result;
+                viewModel.NotificationDelay = result;
         }
 
         private void Button_Mute_Click(object sender, RoutedEventArgs e)
@@ -70,7 +70,7 @@ namespace MystatDesktopWpf.UserControls
             {
                 savedVolume = notificationVolumeSlider.Value;
                 notificationVolumeSlider.Value = 0;
-                
+
                 icon.Kind = PackIconKind.Mute;
             }
             else if (icon.Kind == PackIconKind.Mute)
