+++ MystatDesktopWpf/UserControls/ThemeSettings.xaml
@@ -1,0 +1,46 @@
+﻿<UserControl x:Class="MystatDesktopWpf.UserControls.ThemeSettings"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:MystatDesktopWpf.UserControls"
+             xmlns:controls="clr-namespace:MystatDesktopWpf.UserControls"
+             xmlns:materialDesign="http://materialdesigninxaml.net/winfx/xaml/themes"
+             xmlns:vm="clr-namespace:MystatDesktopWpf.ViewModels"
+             xmlns:converters="clr-namespace:MystatDesktopWpf.Converters"
+             mc:Ignorable="d" 
+             d:DesignHeight="450" d:DesignWidth="800">
+
+    <UserControl.Resources>
+        <converters:ColorToBrushConverter x:Key="ColorToBrushConverter" />
+        <vm:ThemeColorViewModel x:Key="ThemeColorViewModel"/>
+    </UserControl.Resources>
+
+    <materialDesign:Card Padding="8" DataContext="{StaticResource ThemeColorViewModel}">
+        <StackPanel>
+            <TextBlock Text="Настройки темы" HorizontalAlignment="Left"
+                           Style="{DynamicResource MaterialDesignBody1TextBlock}" Margin="0,0,8,8"/>
+            <controls:ThemeSettingsBar Margin="0,0,0,16"/>
+            <Grid>
+                <Grid.ColumnDefinitions>
+                    <ColumnDefinition Width="110"/>
+                    <ColumnDefinition/>
+                </Grid.ColumnDefinitions>
+                <Grid.RowDefinitions>
+                    <RowDefinition Height="Auto"/>
+                    <RowDefinition Height="100"/>
+                </Grid.RowDefinitions>
+                <TextBox Margin="2,0,10,2"
+                       materialDesign:HintAssist.Hint="HEX цвет"
+                       DockPanel.Dock="Top"
+                       Style="{StaticResource MaterialDesignFilledTextBox}"
+                       Text="{Binding Color, ElementName=ColorPicker, UpdateSourceTrigger=PropertyChanged}" />
+                <Rectangle Grid.Row="1" Margin="2,0,10,2" Fill="{Binding Color, ElementName=ColorPicker, Converter={StaticResource ColorToBrushConverter}}" />
+                <materialDesign:ColorPicker x:Name="ColorPicker"
+                                        Grid.Column="1" Grid.RowSpan="2"
+                                        Color="{Binding SelectedColor, Delay=25}" />
+            </Grid>
+        </StackPanel>
+
+    </materialDesign:Card>
+</UserControl>
