+++ MystatDesktopWpf/UserControls/ThemeSettingsBar.xaml
@@ -2,15 +2,15 @@
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-             xmlns:domain="clr-namespace:MystatDesktopWpf.Domain"
+             xmlns:vm="clr-namespace:MystatDesktopWpf.ViewModels"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             xmlns:wpf="clr-namespace:MaterialDesignThemes.Wpf;assembly=MaterialDesignThemes.Wpf"
-             mc:Ignorable="d">
-
+             xmlns:wpf="clr-namespace:MaterialDesignThemes.Wpf;assembly=MaterialDesignThemes.Wpf" 
+             mc:Ignorable="d"
+             d:Background="{DynamicResource MaterialDesignPaper}">
     <UserControl.Resources>
-        <domain:ThemeSettingsViewModel x:Key="ThemeSettingsViewModel"/>
+        <vm:ThemeSettingsViewModel x:Key="ThemeSettingsViewModel"/>
     </UserControl.Resources>
-    
+
     <StackPanel Orientation="Horizontal" DataContext="{StaticResource ThemeSettingsViewModel}">
         <Button Content="Вернуть базовые настройки"/>
