+++ MystatDesktopWpf/ViewModels/ViewModelBase.cs
@@ -2,7 +2,7 @@
 using System.ComponentModel;
 using System.Runtime.CompilerServices;
 
-namespace MystatDesktopWpf.Domain
+namespace MystatDesktopWpf.ViewModels
 {
     public abstract class ViewModelBase : INotifyPropertyChanged
     {
